extracted survey from startpage extracted userId into own action
