+++ actions/questionlist.js
@@ -1,16 +1,16 @@
 import local from './axiosConfigInitial';
 import { error } from './error';
 
-export const USER_ID = 'USER_ID';
+// export const USER_ID = 'USER_ID';
 export const FETCHING_LIST = 'FETCHING_LIST';
 export const BUILD_INITIAL = 'BUILD_INITIAL';
 
-function sendUserId(id) {
-  return {
-    type: USER_ID,
-    id,
-  };
-}
+// function sendUserId(id) {
+//   return {
+//     type: USER_ID,
+//     id,
+//   };
+// }
 function fetchingList() {
   return {
     type: FETCHING_LIST,
@@ -32,9 +32,6 @@ export function questionList() {
       }
       dispatch(buildInitial(resp.data.questions));
       return resp.data;
-    })
-    .then((data) => {
-      return dispatch(sendUserId(data.user.id));
     })
     .catch((err) => {
       return dispatch(error(err));
