+++ public/index.html
@@ -39,6 +39,6 @@
   </head>
   <body>
     <div id="container"></div>
-    <script src="/dist/main.js?5c30951d520eb7196390"></script>
+    <script src="/dist/main.js?9a978872ba9bb7984adc"></script>
   </body>
 </html>
